requirements.txt
Flask==3.1.3
gunicorn==25.1.0
python-dotenv==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
flask-smorest==0.46.2
flask-marshmallow==1.4.0
marshmallow-sqlalchemy==1.4.2
PyJWT==2.11.0
